Rename WindshaftSource
